\name{hflights}
\alias{hflights}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Houston flights data
}
\description{
Dataset based on data available from 'hflights' package. Data provides arrival and departure information from IAH airport.
}
\usage{
hflights
}
\seealso{
For more information see the hflights package.
}