\name{hillmakeR-package}
\alias{hillmakeR-package}
\alias{hillmakeR}
\docType{package}
\title{
hillmakeR
}
\description{
Conduct arrival, departure and occupancy analysis
}
\details{
\tabular{ll}{
Package: \tab hillmakeR\cr
Type: \tab Package\cr
Version: \tab 0.1 \cr
Date: \tab 2013-06-06\cr
License: \tab MIT \cr
}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
David Gilinson
dgilinson@reefpointgroup.com
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
