\name{hillmakeR-package}
\alias{hillmakeR-package}
\alias{hillmakeR}
\docType{package}
\title{
hillmakeR
}
\description{
Conduct arrival, departure and occupancy analysis
}
\details{
\tabular{ll}{
Package: \tab hillmakeR\cr
Type: \tab Package\cr
Version: \tab 0.2 \cr
Date: \tab 2014-07-014\cr
License: \tab MIT \cr
}
The main function in this package is occupancy. See its help for more information.
}
\author{
David Gilinson
dgilinson@reefpointgroup.com
}
\references{
Concept based on software written by MW Isken.
See simular work at http://www.hselab.org/tags/hillmaker.
}
\keyword{ package }
